remove non-error debug print and decompose new files
