+++ lib/services/favorites_service.dart
@@ -186,11 +186,6 @@ class FavoritesService extends ChangeNotifier {
   /// [contentType] specifies whether this is a movie or TV show.
 
   Future<void> addToWatch(Movie movie, {String contentType = 'movie'}) async {
-    debugPrint('📋 [FavoritesService] addToWatch called:');
-    debugPrint('   Movie: ${movie.title} (ID: ${movie.id})');
-    debugPrint('   Movie.contentType: ${movie.contentType}');
-    debugPrint('   Parameter contentType: $contentType');
-
     final toWatch = await getToWatch();
     if (!toWatch.any((m) => m.id == movie.id)) {
       // Ensure contentType is set if not already.
@@ -224,11 +219,6 @@ class FavoritesService extends ChangeNotifier {
   /// [contentType] specifies whether this is a movie or TV show.
 
   Future<void> addToWatched(Movie movie, {String contentType = 'movie'}) async {
-    debugPrint('📋 [FavoritesService] addToWatched called:');
-    debugPrint('   Movie: ${movie.title} (ID: ${movie.id})');
-    debugPrint('   Movie.contentType: ${movie.contentType}');
-    debugPrint('   Parameter contentType: $contentType');
-
     final watched = await getWatched();
     if (!watched.any((m) => m.id == movie.id)) {
       // Ensure contentType is set if not already.
